refactor: extract RelayUtils.swift from AppState
Stage 1 of the AppState god-object split (BACKLOG High Priority).
Moves connectToRelays / publishEventToRelays / fetchEventsFromRelays
plus the _testOnly* DEBUG shims out of AppState.swift into a new
caseless `enum RelayUtils` namespace under Shared/.

Zero behavior change:
- Function bodies and doc comments preserved byte-for-byte.
- Silent error swallow unchanged (deferred to BACKLOG follow-up
  "Better error-message detail in nostrconnect Activity log").
- Actor isolation unchanged (plain async, non-isolated).

Design decisions:
- enum (not struct) namespace, matching Shared/LogExporter.swift.
  Caseless enum is uninstantiable; signals "namespace, not value type".
- Drop _testOnly* wrappers entirely. They existed only because the
  originals were `private`; with extraction the methods are
  internal-by-default and tests call RelayUtils.* directly.
- No Logger added. Stage 1 is structural-only; richer per-relay
  error reporting belongs in the existing BACKLOG follow-up.

Tests:
- AppStateMultiRelayHelpersTests.swift renamed to RelayUtilsTests.swift
  (matches Clave's "what's tested" naming convention).
- 4 tests, same assertions, same inputs. Drop `let appState = AppState()`
  setup line — no longer needed.

Verification:
- xcodebuild test -skip-testing:ClaveUITests on iPhone 17 / iOS 26.4:
  240 passed / 0 failed in both pre-baseline and post-refactor runs.
- AppState.swift -78 lines net (2338 -> 2260).
- New file 71 LOC; pbxproj +4 entries (Clave target only, NSE excluded).

Stages 2-4 of the AppState split (LegacyMigrations, ProfileFetcher,
NostrConnectCoordinator, ProxyClient, AccountManager,
PendingApprovalCoordinator) are out of scope for this PR.
