+++ Clave/AppState.swift
@@ -1582,7 +1582,7 @@ final class AppState {
         }
 
         // Connect to every URI relay in parallel, best-effort.
-        let connectedRelays = await connectToRelays(urls: parsedURI.relays, timeout: 10.0)
+        let connectedRelays = await RelayUtils.connectToRelays(urls: parsedURI.relays, timeout: 10.0)
         defer {
             for relay in connectedRelays { relay.disconnect() }
         }
@@ -1641,7 +1641,7 @@ final class AppState {
             if !handshakeComplete,
                let eventData = connectEvent.toJSON().data(using: .utf8),
                let eventDict = try? JSONSerialization.jsonObject(with: eventData) as? [String: Any] {
-                let acceptedCount = await publishEventToRelays(connectedRelays, event: eventDict)
+                let acceptedCount = await RelayUtils.publishEventToRelays(connectedRelays, event: eventDict)
 
                 if !activityLogged {
                     let success = acceptedCount > 0
@@ -1680,7 +1680,7 @@ final class AppState {
                 "since": now - 10,
                 "limit": 10
             ]
-            let events = await fetchEventsFromRelays(connectedRelays, filter: listenFilter, timeout: 3.0)
+            let events = await RelayUtils.fetchEventsFromRelays(connectedRelays, filter: listenFilter, timeout: 3.0)
             for event in events {
                 guard let eventId = event["id"] as? String, seenEventIds.insert(eventId).inserted else { continue }
                 guard let pubkey = event["pubkey"] as? String,
@@ -2258,81 +2258,4 @@ final class AppState {
         }.resume()
     }
 
-    // MARK: - Multi-relay helpers (nostrconnect handshake)
-
-    /// Connect to multiple relays in parallel, best-effort.
-    /// Returns only the relays that connected successfully within the timeout.
-    /// Failures are silently dropped so one unreachable relay never blocks the others.
-    private func connectToRelays(urls: [String], timeout: TimeInterval) async -> [LightRelay] {
-        await withTaskGroup(of: LightRelay?.self) { group in
-            for url in urls {
-                group.addTask {
-                    let relay = LightRelay(url: url)
-                    do {
-                        try await relay.connect(timeout: timeout)
-                        return relay
-                    } catch {
-                        return nil
-                    }
-                }
-            }
-            var connected: [LightRelay] = []
-            for await maybe in group {
-                if let relay = maybe { connected.append(relay) }
-            }
-            return connected
-        }
-    }
-
-    /// Publish the same event to all connected relays in parallel.
-    /// Returns the number of relays that returned `OK true`.
-    private func publishEventToRelays(_ relays: [LightRelay], event: [String: Any]) async -> Int {
-        await withTaskGroup(of: Bool.self) { group in
-            for relay in relays {
-                group.addTask {
-                    (try? await relay.publishEvent(event: event)) ?? false
-                }
-            }
-            var accepted = 0
-            for await ok in group {
-                if ok { accepted += 1 }
-            }
-            return accepted
-        }
-    }
-
-    /// Fetch events matching the filter from all connected relays in parallel.
-    /// Aggregates results; duplicates by event id are NOT removed (caller should handle).
-    private func fetchEventsFromRelays(
-        _ relays: [LightRelay],
-        filter: [String: Any],
-        timeout: TimeInterval
-    ) async -> [[String: Any]] {
-        await withTaskGroup(of: [[String: Any]].self) { group in
-            for relay in relays {
-                group.addTask {
-                    (try? await relay.fetchEvents(filter: filter, timeout: timeout)) ?? []
-                }
-            }
-            var all: [[String: Any]] = []
-            for await events in group {
-                all.append(contentsOf: events)
-            }
-            return all
-        }
-    }
-
-    // MARK: - Test-only shims
-
-    #if DEBUG
-    func _testOnlyConnectToRelays(urls: [String], timeout: TimeInterval) async -> [LightRelay] {
-        await connectToRelays(urls: urls, timeout: timeout)
-    }
-    func _testOnlyPublishEventToRelays(_ relays: [LightRelay], event: [String: Any]) async -> Int {
-        await publishEventToRelays(relays, event: event)
-    }
-    func _testOnlyFetchEventsFromRelays(_ relays: [LightRelay], filter: [String: Any], timeout: TimeInterval) async -> [[String: Any]] {
-        await fetchEventsFromRelays(relays, filter: filter, timeout: timeout)
-    }
-    #endif
 }
